Task 2.2C/requirements.txt
asyncio
bleak
pandas
google-cloud-bigquery
pyarrow
numpy
